backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.10"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.13",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "alembic==1.16.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "async-timeout==4.0.3",
    "attrs==25.3.0",
    "banks==2.1.3",
    "bcrypt==3.2.0",
    "beautifulsoup4==4.13.4",
    "boto3==1.39.3",
    "botocore==1.39.3",
    "bs4==0.0.2",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "cryptography==44.0.3",
    "cssselect==1.3.0",
    "dataclasses-json==0.6.7",
    "dateparser==1.2.1",
    "deprecated==1.2.18",
    "dirtyjson==1.0.8",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "frozenlist==1.7.0",
    "fsspec==2025.3.2",
    "google-auth==2.40.3",
    "google-genai==1.23.0",
    "greenlet==3.1.1",
    "griffe==1.7.3",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langsmith==0.4.4",
    "lxml==5.3.2",
    "lxml-html-clean==0.4.2",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.6.2",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numpy==2.2.4",
    "openai==1.82.0",
    "orjson==3.10.18",
    "packaging==24.2",
    "pandas==2.2.3",
    "passlib==1.7.4",
    "pgvector==0.4.0",
    "pillow==11.2.1",
    "pip-autoremove==0.10.0",
    "platformdirs==4.3.8",
    "propcache==0.3.2",
    "psutil==7.0.0",
    "psycopg2-binary",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings>=2.12.0",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pypdf==5.6.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "readability-lxml==0.8.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "rich-toolkit==0.14.1",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.40",
    "starlette==0.46.1",
    "striprtf==0.0.26",
    "tenacity==8.5.0",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tomli==2.2.1",
    "tqdm==4.67.1",
    "transfer==0.13.0",
    "typer==0.15.2",
    "typing-extensions==4.13.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.0",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
